+++ app/services/cache/clean.rb
@@ -4,12 +4,17 @@ class Cache::Clean
   class << self
     def call
       Rails.logger.info('Cleaning cache...')
+      delete_control_flag
       delete_version_cache
       delete_years_tracked_cache
       Rails.logger.info('Cache cleaned')
     end
 
     private
+
+    def delete_control_flag
+      Rails.cache.delete('cache_jobs_scheduled')
+    end
 
     def delete_version_cache
       Rails.cache.delete(CheckAppVersion::VERSION_CACHE_KEY)