+++ config/environment.rb
@@ -6,8 +6,11 @@ require_relative 'application'
 # Initialize the Rails application.
 Rails.application.initialize!
 
-# Clear the cache
-Cache::CleaningJob.perform_later
+# Use an atomic operation to ensure one-time execution
+if defined?(Rails::Server) && Rails.cache.write('cache_jobs_scheduled', true, unless_exist: true)
+  # Clear the cache
+  Cache::CleaningJob.perform_later
 
-# Preheat the cache
-Cache::PreheatingJob.perform_later
+  # Preheat the cache
+  Cache::PreheatingJob.perform_later
+end